Reformatted error/success messages
Noticed an issue with the 'download failed' message not having the correct spacing, went back to the old version with the space in the Timestamp.
Would like to make the tkinter error messages and lose the command prompt window, but haven't gotten around to it yet. Didn't make this into a new version of the .exe, don't feel like it's a big enough change to warrant it.